Organize RPC namespace folders
This organized RPC namespace folders into types, serializers, and utils.
Before the items were randomly assorted into other folders of different
names.
