+++ ironfish/src/rpc/routes/peer/index.ts
@@ -7,3 +7,4 @@ export * from './getPeer'
 export * from './getBannedPeers'
 export * from './getPeerMessages'
 export * from './addPeer'
+export * from './types'
