+++ ironfish/src/rpc/routes/wallet/burnAsset.test.ts
@@ -11,7 +11,7 @@ import {
 } from '../../../testUtilities'
 import { createRouteTest } from '../../../testUtilities/routeTest'
 import { CurrencyUtils } from '../../../utils'
-import { serializeRpcWalletTransaction } from './utils'
+import { serializeRpcWalletTransaction } from './serializers'
 
 describe('Route wallet/burnAsset', () => {
   const routeTest = createRouteTest(true)
