+++ ironfish/src/rpc/routes/wallet/getAsset.ts
@@ -5,10 +5,10 @@ import { ASSET_ID_LENGTH } from '@ironfish/rust-nodejs'
 import * as yup from 'yup'
 import { CurrencyUtils } from '../../../utils'
 import { RpcNotFoundError, RpcValidationError } from '../../adapters'
+import { RpcAsset, RpcAssetSchema } from '../chain/types'
 import { ApiNamespace } from '../namespaces'
 import { routes } from '../router'
 import { AssertHasRpcContext } from '../rpcContext'
-import { RpcAsset, RpcAssetSchema } from '../types'
 import { getAccount } from './utils'
 
 export type GetWalletAssetRequest = {
