+++ ironfish/src/rpc/routes/wallet/getAssets.ts
@@ -3,10 +3,10 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import * as yup from 'yup'
 import { CurrencyUtils } from '../../../utils'
+import { RpcAsset, RpcAssetSchema } from '../chain'
 import { ApiNamespace } from '../namespaces'
 import { routes } from '../router'
 import { AssertHasRpcContext } from '../rpcContext'
-import { RpcAsset, RpcAssetSchema } from '../types'
 import { getAccount } from './utils'
 
 export type GetAssetsRequest = {
