+++ ironfish/src/wallet/exporter/encoders/json.ts
@@ -3,8 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { generateKeyFromPrivateKey } from '@ironfish/rust-nodejs'
 import { Assert } from '../../../assert'
+import { deserializeRpcAccountMultisigKeys } from '../../../rpc/routes/wallet/serializers'
 import { RpcAccountImport } from '../../../rpc/routes/wallet/types'
-import { deserializeRpcAccountMultisigKeys } from '../../../rpc/routes/wallet/utils'
 import { validateAccount } from '../../validator'
 import { AccountImport } from '../accountImport'
 import { AccountDecodingOptions, AccountEncoder, DecodeFailed } from '../encoder'
